Remove the redundant bootstrap.py module. Move python code out of __init__.py scripts into their own modules.
