--- packages/bootstrap/extra/dcos_internal_utils/__init__.py
@@ -1,2 +1,0 @@
-from . import bootstrap  # noqa
-from . import exhibitor  # noqa
