+++ packages/bootstrap/extra/dcos_internal_utils/bootstrap.py
@@ -10,7 +10,7 @@ from kazoo.client import KazooClient
 from kazoo.retry import KazooRetry
 from kazoo.security import ACL, ANYONE_ID_UNSAFE, Permissions
 
-from .. import utils
+from dcos_internal_utils import utils
 
 log = logging.getLogger(__name__)
 
@@ -65,11 +65,13 @@ class Bootstrapper(object):
                 return zkid
 
     def generate_oauth_secret(self, path):
+        log.info('Generating oauth secret at {}'.format(path))
         possible_auth_token = ''.join(random.choice(string.ascii_letters) for _ in range(64))
         self.zk.ensure_path('/dcos', ANYONE_ALL)
         consensus_auth_token = self._consensus('/dcos/auth-token-secret',
                                                possible_auth_token.encode('ascii'), ANYONE_READ)
         _write_file(path, consensus_auth_token, 0o600, 'dcos_oauth')
+        return consensus_auth_token
 
     def _consensus(self, path, value, acl=None):
         if value is not None:
@@ -87,14 +89,18 @@ class Bootstrapper(object):
 
 def _write_file(path, data, mode, owner='root'):
     dirpath = os.path.dirname(os.path.abspath(path))
+    log.info('Opening {} for locking'.format(dirpath))
     with utils.Directory(dirpath) as d:
+        log.info('Taking exclusive lock on {}'.format(dirpath))
         with d.lock():
             umask_original = os.umask(0)
             try:
                 flags = os.O_WRONLY | os.O_CREAT | os.O_TRUNC
                 log.info('Writing {} with mode {:o}'.format(path, mode))
-                with os.fdopen(os.open(path, flags, mode), 'wb') as f:
+                tmppath = path + '.tmp'
+                with os.fdopen(os.open(tmppath, flags, mode), 'wb') as f:
                     f.write(data)
+                os.rename(tmppath, path)
                 user = pwd.getpwnam(owner)
                 os.chown(path, user.pw_uid, user.pw_gid)
             finally:
