+++ packages/bootstrap/extra/setup.py
@@ -11,9 +11,6 @@ setup(
     install_requires=requires,
     packages=[
         'dcos_internal_utils',
-        'dcos_internal_utils.bootstrap',
-        'dcos_internal_utils.exhibitor',
-        'dcos_internal_utils.utils',
     ],
     version='0.0.1',
     description='DC/OS Internal Utilities Library',
@@ -24,6 +21,6 @@ setup(
         'Programming Language :: Python :: 3.4'],
     entry_points={
         'console_scripts': [
-            'bootstrap=dcos_internal_utils.bootstrap.cli:main'
+            'bootstrap=dcos_internal_utils.cli:main'
         ]
     })
